library: Fix a bug where a duplicated album is labeled various artists
When checking for various artists, we don't check the actual artist
tags. It's possible that a user has two copies of some song, or even an
entire album, in their music library. In this case we would wrongly
label the album as various artists.

Note this doesn't fix listing the songs in the album twice. We believe
duplicates are an error in the library.