info: don't crash when a stream has no tracks
update_album() would cause Sonata to crash when playing a stream with
artist and album metadata. The call to album_return_artist_and_tracks()
would rightly return None and None for the artist and tracks because the
stream is non-local. To avoid this reorganize the checks to make sure
there are tracks to display in the album info. Otherwise just display
"Album info not found." and be done with it.