album-covers: replace amazon.com with rhapsody.com cover fetching
As of August 15th, 2009 amazon.com requires all requests sent to their
server to be signed with the private key corresponding to the public key
being used. Since we're an open source project we cannot do this without
exposing the private key. Therefore switch to using rhapsody.com's
webservice.

Backported for 1.6.2