main: fix centering of current song in current playlist
Previously, the current song would be centered in the current playlist
if the file is different than the previous song's file. It's possible
for duplicate songs to be present in the current playlist, and thus
this test would fail. Instead of using the file, use the song's
id which should be unique within the playlist.