Fix crashes when files cannot be opened for reading/writing
Previously, we wouldn't catch IOError's and OSError's when a file
couldn't be opened for reading or writing. Fix this by surrounding those
locations in try-except clauses and use the 'with' syntax to properly
close files when errors occur.

This resolves issues for users with a read-only home or otherwise
read-only directories. Sadly, the artwork code downloads images directly
to disk, and if the file cannot be opened there is no artwork shown.
Fixing this is a larger issue we can resolve later.

Fixes: #16355