library: don't list albums with multiple discs multiple times
If a user has a music directory where an album has multiple directories
for each disc, we would wrongly list that album multiple times in the
album view. This is because when we remove duplicate album listings, the
differing paths lead us to believe the albums are not the same. Fix this
by ignoring the path of the music files so when we remove the duplicates
we are left with just one album instead of multiple ones.

Fixes: #16368
Tested-by: Reuben Thomas <rrt@sc3d.org>