systray: remove the useless callback on the popup-menu event for gtk.StatusIcon
Kirill Lashuk (KL-7) on Github discovered that this event did the same thing (in
the end) that button-press-event, which resulted in the context menu to be
displayed two times when using gtk.StatusIcon (there was a little flick).

I removed the registration of the callback on the popup-menu event. Since the
associated callback was only use by gtk.StatusIcon, to do the same thing as the
button-press-event callback, I also removed it, and simplified a little bit the
'tray API'.