ui-connection: display the 'offline' UI if mpd doesn't run at Sonata startup
Kirill Lashuk (KL-7) from Github discovered that the previous fix on this branch
made Sonata display the 'online' UI if it was launched while mpd wasn't running.

This is due to the missing call to Base.handle_change_conn in Base.iterate if
'self.conn == self.prevconn', and the fact that the default behavior is to be
'connected' by default.