lyrics: ask the lyrics plugin to force the retrieval of new lyrics
Previously, this was done in the main module, which had to guess a bit how the
lyrics system was working: the lyrics file had to be deleted before fetching the
new lyrics, and the call to the function which located the lyrics file was
buggy.
This fixes the following error on fetching a song throught the "Search" button:

    Traceback (most recent call last):
      File ".../sonata/info.py", line 221, in on_link_click
        self.on_link_click_cb(linktype)
      File ".../sonata/main.py", line 3242, in on_link_click
        self.on_lyrics_search(None)
      File ".../sonata/main.py", line 2214, in on_lyrics_search
        consts.LYRICS_LOCATION_HOME)
      File ".../sonata/info.py", line 534, in target_lyrics_filename
        os.path.join(*file_map[lyrics_loc]))
      File "/usr/lib/python2.6/posixpath.py", line 65, in join
        if b.startswith('/'):
    AttributeError: 'NoneType' object has no attribute 'startswith'

Now, we let the lyrics system handle all of this, and we ask it to force the
fetching, even if a lyrics file already existed.