launcher: make the main launcher cleaner
The launch script has been moved to the 'sonata' package, as the rest of the
code, an put into a dedicated function, that is launch now by a very simple
script at the top level. It will allow to 'import sonata' from the source
repository and having the right behavior ('sonata' will by the package not the
top level script).

Additionnaly, Sonata now depends on setuptools/distribute to generate
automatically a good startup script. Both the commited one and the one created
by setuptools can be used.