Convert info page into GtkBuilder-built UI
There are some holdovers here, as with the about, where it may be
more maintainable to build in the rest of the UI from python.

That's usually where you see a tight loop over a hand-built array of
strings.  For both coding and localizing, it's probably easier to refer
to the array than a tangle of markup.