py3: open files using UTF-8, no matter what the default system encoding is
Otherwise, this relies on the locale.getpreferredencoding(), which can't be
overridden.

There are still some cases where Sonata crashes on Unicode errors, for
example if the filesystem contains files with non-ascii characters which
Sonata tries to access somehow (like os.path.exists()). I don't see how
to fix this beside encoding file names to UTF-8 each time we access the
filesystem. Which I'm not going to do since there *must* be a better
solution.

This should fix #25 (*prey*)