lyrics: new, simpler, lyricswiki plugin
The new plugin comes from Anton Lashkov, adapted to Python 3 by me.

A side effect is that it's not possible to directly edit the lyrics from
Sonata anymore, due to the fact that the plugin doesn't export the "edit"
functionality anymore. This might come back later on, if really needed.

Co-Edited-By: Jonathan Ballet <jon@multani.info>