covers: fix size of covers in the search remote covers dialog
For some reasons, it was working a long time ago, but didn't since a
while. I couldn't bisect the changeset, too many non-working changesets
during the Py3/Gtk3 migration branch :/

"162" is an arbitrary value, which is supposed to be the size of the
image plus the size of the "spiner" on the left.