current: fix and simplify searching songs in the current tab.
Searching songs in the current tab now relies on the Gtk.TreeModelFilter
model, which (as the name implies) does filtering on the model.
When a search is performed, it updates the pattern to search and filter
on the go the rows given to the view. The view always deals with this
filtering model, although it doesn't filter all the time.

This is a huge improvement over the custom filtering method used until now,
which consisted of creating a brand new model with filtered values, dealing
with stringly typed variables (?!), messing with the model used by the
view, with some caching, optimization, threads and lock on the way.
Although the search is not faster, the code is much, much, much simpler to
read and understand.
Some cleans are still doable, although with some optimizations which will
be done later.