lyrics: follow redirect pages on LyricsWiki
Some pages in LyricsWiki actually redirect to another page to find the actual lyrics. This is the
case for example for "The Clash: Police & Thieves" which contains a redirect marker to
"The Clash: Police and Thieves". If we find this kind of marker, we redo the search with the new
informations.
With a special crafted page, it's possible to stuck Sonata in a forever-loop, so we also try not to
get stuck.