thread: MPD connections are not in a thread anymore
* mpd.MPDClient() is not thread-safe, per python-mpd2 documentation
* there were race conditions if the connection was too slow
* this thread was doing too much things outside of the main thread
* long commands were blocking Sonata anyway due to bad threading management

It works as good (and as bad) as before. It makes sense to have the
connection to MPD not blocking the UI, but it has to be done correctly,
and I prefer to remove all the bad threading management now, and
refactor this part later if it causes problem and we find a good way to
do it.