artwork: fix art cache loading/saving
The art cache is a dictionary which is simply dumped and loaded using eval()
(!), but the SongRecord object, used as a cache key was dumped as a complex
structure which couldn't be loaded back (something like "<SongRecord...>".
So, even if the cache was saved, it could never be loaded afterwards.

Now, the key is a simple tuple, which makes everything work.