lyrics: better separation between threads when updating lyrics
The lyrics plugins are launched in a dedicated thread which used to update
the GUI directly without using the main thread to do it.
This is not really cool with Gtk, and so the thread function is now
completely detached from the rest of the "Info" tab code, and communicate
with it via a dedicated method which schedule modifications of the GUI
through the main thread.

Fix: #56