Uniform handling of locale initialization.
Provide uniform translation for both GObject / Glade strings
(initialized by module locale) and Python gettext strings to
prevent a mix of languages in UI strings.

If a locale is provided with sonata but not generated for the C
library, Python strings will be translated, but strings in Glade
dialogs will not be translated. So if the initial locale.setlocale
call fails, don't point Python gettext to the .mo files -- this way
all strings will be left untranslated.