Avoid problematic uses of 'is'
The 'is' operator compares by object identity, which is not usually
what you want unless comparing with singleton objects like None.
Python 3.8 warns about suspicious uses of 'is' during byte-compilation.

As an implementation detail, CPython interns small integers, so
"1 + 1 is 2" is likely to be true, although this is not an API guarantee.

However, "field is 'comment'" is unlikely to be true, because CPython
allocates separate string objects for each instance of the same long
string.

Signed-off-by: Simon McVittie <smcv@debian.org>