launcher: Assume the GDK lock is already held in GtkApplication::activate (#165)
GtkApplication is documented to acquire the GDK lock when we receive
::activate from another process. The GDK lock is not guaranteed to be
a recursive lock, and if it is not, we will deadlock if we try to
acquire it when already acquired.

This particular application does not seem to emit the activate signal
manually, but if it does at some point in future, it will need to
make sure to emit the signal in a context where it would be safe to
call GTK functions.

See also docs/guide/threading.rst in pygobject.

Resolves: https://github.com/multani/sonata/issues/164

Signed-off-by: Simon McVittie <smcv@debian.org>